Reupload
This project is based on setting up a client to cache to server setup. The client checks the cache for a file. If the cache has the file, it just send it. If it doesn't, the cache gets the file from the server. This was meant to be run on different machines, we used virtual machines. This has no functional use, but was an interesting project nonetheless. 

There was also a client that would attempt to get a large file, however the large file was an ISO of Soulcalibur 2 or 3 for the PS2 and that file shouldn't be hosted on github for both its size and because I don't have the rights to that in any way.

Please ignore/enjoy my somewhat childish commentary.